requirements.txt
streamlit==1.28.1
openai==1.3.7
bcrypt==4.1.2
python-dotenv==1.0.0
reportlab==4.0.7
pandas==2.1.4
opencv-python==4.8.1.78
numpy==1.24.3
PyMuPDF==1.23.8
python-docx==1.1.0
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
python-magic-bin==0.4.14
Pillow==10.1.0
